+++ src/search/index.ts
@@ -1,0 +1,6 @@
+export * from './types.js';
+export * from './utils.js';
+export * from './tag-search.js';
+export * from './status-search.js';
+export * from './block-ref-search.js';
+export * from './hierarchy-search.js';